created router for products
